--- src/main/java/com/bit/joe/shoppingmall/controller/AdminController.java
@@ -1,14 +1,0 @@
-package com.bit.joe.shoppingmall.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/admin")
-public class AdminController {
-    @GetMapping
-    public String index() {
-        return "thymeleaf/admin/insight";
-    }
-}
