+++ src/main/java/com/bit/joe/shoppingmall/controller/testControllers/AdminController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.controller.testControllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/admin/inventory")
-public class InventoryController {
+@RequestMapping("/admin")
+public class AdminController {
     @GetMapping
     public String index() {
-        return "thymeleaf/admin/inventory";
+        return "thymeleaf/admin/insight";
     }
 }
