+++ src/main/java/com/bit/joe/shoppingmall/controller/testControllers/CartHtmlController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.controller.testControllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/history")
-public class HistoryController {
+@RequestMapping("/cart")
+public class CartHtmlController {
     @GetMapping
-    public String history() {
-        return "thymeleaf/product/history";
+    public String cart() {
+        return "thymeleaf/product/cart";
     }
 }
