+++ src/main/java/com/bit/joe/shoppingmall/controller/testControllers/HistoryController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.controller.testControllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/cart")
-public class CartHtmlController {
+@RequestMapping("/history")
+public class HistoryController {
     @GetMapping
-    public String cart() {
-        return "thymeleaf/product/cart";
+    public String history() {
+        return "thymeleaf/product/history";
     }
 }
