+++ src/main/java/com/bit/joe/shoppingmall/controller/testControllers/HomeController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.controller.testControllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/product")
-public class ProductdetailController {
+@RequestMapping("/")
+public class HomeController {
     @GetMapping
-    public String productdetail() {
-        return "thymeleaf/product/product-detail";
+    public String index() {
+        return "thymeleaf/index";
     }
 }
