+++ src/main/java/com/bit/joe/shoppingmall/controller/testControllers/MemberController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.controller.testControllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/")
-public class HomeController {
+@RequestMapping("/admin/member")
+public class MemberController {
     @GetMapping
     public String index() {
-        return "thymeleaf/index";
+        return "thymeleaf/admin/member";
     }
 }
