--- src/main/resources/templates/thymeleaf/index.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="en">
-<head>
-    <title>Home Page</title>
-</head>
-<body>
-</body>
-</html>
