+++ src/main/resources/templates/thymeleaf/user/customuserinterface.html
